function/requirements.txt
azure-functions
azure-servicebus
psycopg2-binary
sendgrid
python-dotenv
